Corrections for all previous commits
