+++ src/foam/dao/DAO.js
@@ -27,7 +27,7 @@ foam.INTERFACE({
       async: true,
       type: 'FObject',
       swiftThrows: true,
-      documentation: '',
+      documentation: 'Inserts a new object or updates an existing one. When the object is stored successfully, the promise resolves with the newly added object.',
       args: [
         {
           name: 'obj',
@@ -39,7 +39,6 @@ foam.INTERFACE({
       name: 'put_',
       async: true,
       type: 'FObject',
-      documentation: '',
       swiftThrows: true,
       args: [
         {
@@ -57,7 +56,9 @@ foam.INTERFACE({
       async: true,
       type: 'FObject',
       swiftThrows: true,
-      documentation: '',
+      documentation: `Deletes a single object from the DAO.
+                      NB: Trying to remove an object which does not exist is not an error. remove() only rejects if it fails to communicate with the backend.
+      `,
       args: [
         {
           name: 'obj',
@@ -70,7 +71,6 @@ foam.INTERFACE({
       async: true,
       type: 'FObject',
       swiftThrows: true,
-      documentation: '',
       args: [
         {
           name: 'x',
@@ -87,7 +87,10 @@ foam.INTERFACE({
       async: true,
       type: 'FObject',
       swiftThrows: true,
-      documentation: 'Retrieves a single object from the DAO, whose id is equal to the parameter id. If the object is found, the promise resolves with the object. If the object is not found, it rejects with a foam.dao.ObjectNotFoundException.',
+      documentation: `Retrieves a single object from the DAO, whose id is equal to the parameter id or if the given predicate evaluates to true.
+      If the object is found, the promise resolves with the object.
+      If the object is not found, it returns null.
+      `,
       args: [
         {
           name: 'id',
@@ -100,7 +103,6 @@ foam.INTERFACE({
       async: true,
       type: 'FObject',
       swiftThrows: true,
-      documentation: 'Retrieves a single object from the DAO, whose id is equal to the parameter id. If the object is found, the promise resolves with the object. If the object is not found, it rejects with a foam.dao.ObjectNotFoundException.',
       args: [
         {
           name: 'x',
@@ -117,8 +119,9 @@ foam.INTERFACE({
       async: true,
       type: 'foam.dao.Sink',
       swiftThrows: true,
-      documentation: '',
-      args: [
+      documentation: `The primary way to read objects from a DAO is select(sink) that will retrieve a collection of results and will send them (callback) to the sink. A simple select(sink) returns all record in the DAO.
+        If you don\'t specify a Sink when calling select(), a foam.dao.ArraySink will be created by default and passed to the resolved Promise:
+      `,      args: [
         {
           name: 'sink',
           type: 'foam.dao.Sink',
@@ -129,7 +132,6 @@ foam.INTERFACE({
       name: 'select_',
       async: true,
       type: 'foam.dao.Sink',
-      documentation: '',
       swiftThrows: true,
       args: [
         {
@@ -167,14 +169,15 @@ foam.INTERFACE({
       name: 'removeAll',
       async: true,
       swiftThrows: true,
-      documentation: '',
+      documentation: `removeAll() is very similar to select(), with the obvious exception that it removes all matching entries from the DAO instead of returning them.
+                      Be careful! myDAO.removeAll() without any filtering will delete every entry.
+      `,
       args: [ ]
     },
     {
       name: 'removeAll_',
       swiftThrows: true,
       async: true,
-      documentation: '',
       args: [
         {
           name: 'x',
